refactor: rename atlas-blueprints folder to headless-blueprints
